Added announce command
Presenting a new feature for Gamebot:
Announcements!
So you can run this command in any channel but it is recommended you run it in here and only with permission.
The syntax for this command is as follows: /announce <Channel> <Mention Everyone?> <Text to announce>
The commmand won't be deleted which is why you should keep it to this channel or botspam

This message was in fact sent using this command. The blue bar on the left is my name color and my name and profile pic are visable at the top